+++ lib/src/widgets/async_fab/async_fab.dart
@@ -1,5 +1,5 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/change_notifier.dart';
 import 'package:flutter_async/src/async_controller.dart';
 import 'package:flutter_async/src/widgets/async_state.dart';